AddList repaired Redo/Undo stack fixed Code allocation updated

+++ src/storage/DatabaseAccessor.java
@@ -3,6 +3,7 @@ package storage;
 import backend.ListElement;
 import controls.Controller;
 
+import javax.swing.*;
 import java.io.*;
 
 public class DatabaseAccessor<T> {
@@ -13,16 +14,26 @@ public class DatabaseAccessor<T> {
 
     }
 
-    public void addList(String name) {
+    public boolean addList(String name) {
         // Create empty csv-file
         // Erstelle eine neue leere CSV-Datei im Projektordner
+
         File file = new File(source + name + ".csv");
+        if (file.exists()) {
+            int ans = JOptionPane.showConfirmDialog(null, "Es gibt bereits eine Datei mit diesem Namen.\r\nDatei überschreiben?", "Datei existiert bereits", JOptionPane.YES_NO_OPTION, JOptionPane.QUESTION_MESSAGE);
+
+            if (ans == JOptionPane.NO_OPTION || ans == JOptionPane.CLOSED_OPTION) {
+                return false;
+            }
+        }
 
         try (BufferedWriter writer = new BufferedWriter(new FileWriter(file))) {
-            writer.write("");
+            writer.write("data");
             System.out.println("CSV-Datei erstellt: " + file.getAbsolutePath());
+            return true;
         } catch (IOException e) {
             System.err.println("Fehler beim Erstellen der Liste: " + e.getMessage());
+            return false;
         }
     }
